+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/confluent/client/ConfluentSchemaRegistryClient.scala
@@ -66,7 +66,6 @@ class DefaultConfluentSchemaRegistryClient(override val client: CSchemaRegistryC
     SchemaWithMetadata(AvroUtils.adjustParsedSchema(rawSchema, config), id)
   }
 
-  override def getLatestSchemaId(topic: String, isKey: Boolean): Validated[SchemaRegistryError, SchemaId] = getLatestFreshSchema(topic, isKey).map(_.id)
 }
 
 object ConfluentSchemaRegistryClient {